refactor(create-plugin): name export create plugin and utils
- Update all plugin createPlugin imports
- Use mapSchema util on server plugins

BREAKING CHANGE: Mapping schema's must use mapSchema utility
