+++ packages/plugins/src/client/action.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { fetchGetById } from './fetch.js'
 import { dataGetValue } from './data.js'
 import { operatorCompare, operatorEval } from './operator.js'
