+++ packages/plugins/src/client/component.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import {
   dataUnsafeSetValue,
   dataGetValue,
