+++ packages/plugins/src/client/editor.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { dataGetSchema } from './data.js'
 
 /**
