+++ packages/plugins/src/client/event.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { actionDispatch, dataGetValue } from './index.js'
 
 /**
