+++ packages/plugins/src/client/form.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { dataGenerateId } from './data.js'
 
 /**
