+++ packages/plugins/src/client/icon.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { dataGetValue, dataSetValue } from './data.js'
 
 let iconifyAPIUrl = 'https://api.iconify.design'
