+++ packages/plugins/src/client/list.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { actionDispatch, operatorCompare, operatorEval } from './index.js'
 
 /**
