+++ packages/plugins/src/client/operator.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 
 /**
  * Two or one values to run an operator on
