+++ packages/plugins/src/client/regex.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 
 /**
  * @typedef {Object} RegexPattern
