+++ packages/plugins/src/client/string.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 
 /**
  * @typedef {Object} StringReplace
