+++ packages/plugins/src/client/token.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { dataGetValue } from './data.js'
 
 const tokenProcess = {}
