+++ packages/plugins/src/client/variable.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin } from '@dooksa/create-plugin'
 import { dataGetValue, dataSetValue } from './data.js'
 import { generateId, getValue } from '@dooksa/utils'
