+++ packages/plugins/src/server/component.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin, mapSchema } from '@dooksa/create-plugin'
 import { component as componentClient, dataGenerateId, pagePathToId } from '#client'
 import { databaseDeleteValue, databaseGetValue, databaseSeed, httpSetRoute } from '#server'
 import { components as defaultComponents } from '@dooksa/components'
@@ -6,7 +6,7 @@ import { pageCreate } from '#server'
 import { createDataValue } from '#utils'
 
 export const component = createPlugin('component', {
-  schema: componentClient.schema,
+  schema: { ...mapSchema(componentClient) },
   /**
    * @param {Object} options
    * @param {import('@dooksa/create-component').Component[]} [options.components]
