+++ packages/plugins/src/server/event.js
@@ -1,9 +1,9 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin, mapSchema } from '@dooksa/create-plugin'
 import { event as eventClient } from '../client/index.js'
 import { databaseSeed } from './database.js'
 
 export const event = createPlugin('event', {
-  schema: { ...eventClient.schema },
+  schema: { ...mapSchema(eventClient) },
   setup () {
     databaseSeed('event-listeners')
     databaseSeed('event-handlers')
