+++ packages/plugins/src/server/metadata.js
@@ -1,4 +1,4 @@
-import createPlugin from '@dooksa/create-plugin'
+import { createPlugin, mapSchema } from '@dooksa/create-plugin'
 import { dataSetValue, metadata as metadataClient } from '../client/index.js'
 import { databaseSeed, databaseGetValue } from './database.js'
 import { httpSetRoute } from './http.js'
@@ -9,7 +9,7 @@ import { httpSetRoute } from './http.js'
 
 export const metadata = createPlugin('metadata', {
   schema: {
-    ...metadataClient.schema
+    ...mapSchema(metadataClient)
   },
   /**
    * @param {Object} param
